requirements.txt
aiocache==0.12.3
alembic==1.14.0
asyncpg==0.30.0
fastapi==0.115.6
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.37
starlette==0.41.3
web3==7.6.1
pydantic_settings==2.7.1
uvicorn==0.34.0